MONOLITH is now in a class! that inherits from QOject
reason for doing this:
The main reason for doing this is so that we can create signals and
slots for the GUI.
Next step. Connect a UI button to addParticle.

Embedding the Binaries in classes is also good so we can run instances
of them. next step. try running particle demo and MONOLITH both at the
same time.